Add support for _id export (#12)
Prior to this change, we had not exported "_id" in our get endpoints which made querying by unique values difficult. This change adds a "_id" result.

By default, our framework, Rocket, adds the following format:

```json
{
   "_id": 
   {
      "$oid": "mycustomhex"
   }
}
```

Though we can further process this and remove it, I think it is helpful to keep it as-is to discern between normal strings and OID strings. Right now "_id" is the only key that uses this formatting.